Add auto_stick_play parameter to ROS node
When starting the tm_driver we can use the `stick_play` value to control
whether or not the robot will press the stick play button.

The ROS node however does not use that feature, always using the default
`true` value. This means that starting the driver will start the default
program on the robot if it is in Auto mode. This is a potential safety
issue.

Added a auto_stick_play node parameter to allow overriding that value,
allowing the node to start the driver to start with `stick_play` to
false.